requirements.txt
openai>=1.40.0
playwright>=1.44.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
rich>=13.7.0
 